+++ CsprojToAsmdef.Build/Build.cs
@@ -1,6 +1,5 @@
 using Nuke.Common;
 using Nuke.Common.CI;
-using Nuke.Common.Execution;
 using Nuke.Common.ProjectModel;
 using Nuke.Common.Tools.DotNet;
 using System;
@@ -9,24 +8,26 @@ using System.Linq;
 using System.Text.Json;
 using static Nuke.Common.Tools.DotNet.DotNetTasks;
 
-[CheckBuildProjectConfigurations]
 [ShutdownDotNetAfterServerBuild]
 partial class Build : NukeBuild
 {
-    public static int Main() => Execute<Build>(x => x.Compile);
-
     [Parameter("Configuration to build - Default is 'Debug' (local) or 'Release' (server)")]
     readonly Configuration Configuration = IsLocalBuild ? Configuration.Debug : Configuration.Release;
 
-    [Parameter("Should be true for continuous integration builds")] readonly bool CiBuild;
+    [Parameter("Should be true for continuous integration builds")]
+    readonly bool CiBuild;
 
     [Solution] readonly Solution Solution = default!;
+
+    string PackageDirectory => Path.Combine(Solution.Directory, "CsprojToAsmdef", "Assets", "CsprojToAsmdef");
 
     string CurrentVersion
     {
         get
         {
-            var packagePath = Path.Combine(Solution.Directory, "CsprojToAsmdef", "Assets", "CsprojToAsmdef", "package.json");
+            var packageDirectory = PackageDirectory;
+
+            var packagePath = Path.Combine(packageDirectory, "package.json");
 
             if (!File.Exists(packagePath)) throw new InvalidOperationException($"package.json does not exist at path: {packagePath}");
 
@@ -46,7 +47,7 @@ partial class Build : NukeBuild
             DotNetRestore(s => s
                 .SetProjectFile(CliProject)));
 
-    Target Compile => _ => _
+    Target BuildCli => _ => _
         .DependsOn(Restore)
         .Executes(() =>
             DotNetBuild(s => s
@@ -54,5 +55,19 @@ partial class Build : NukeBuild
                 .SetConfiguration(Configuration)
                 .SetVersion(CurrentVersion)
                 .SetContinuousIntegrationBuild(CiBuild)
+                .EnableNoIncremental()
                 .EnableNoRestore()));
+
+    Target BuildAll => _ => _
+        .DependsOn(InstallCli)
+        .Executes(() =>
+            DotNetBuild(s => s
+                .SetProjectFile(Solution)
+                .SetConfiguration(Configuration)
+                .SetVersion(CurrentVersion)
+                .SetContinuousIntegrationBuild(CiBuild)
+                .EnableNoIncremental()
+                .EnableNoRestore()));
+
+    public static int Main() => Execute<Build>(x => x.BuildAll);
 }