+++ CsprojToAsmdef.Build/Ci.cs
@@ -1,0 +1,11 @@
+﻿using Nuke.Common;
+
+partial class Build
+{
+    Target RunCi => _ => _
+        .DependsOn(CheckFormatting)
+        .DependsOn(InstallCli)
+        .Executes(() =>
+        {
+        });
+}