+++ CsprojToAsmdef.Cli/Domain/Asmdef.cs
@@ -1,61 +1,43 @@
-﻿using System.Collections.Generic;
+﻿using Microsoft.Build.Evaluation;
+using MoreLinq.Extensions;
+using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Linq;
 using System.Text.Json;
-using System.Text.Json.Serialization;
-using Microsoft.Build.Evaluation;
 
 namespace CsprojToAsmdef.Cli.Domain
 {
-    [SuppressMessage("ReSharper", "UnusedAutoPropertyAccessor.Global")]
-    [SuppressMessage("CodeQuality", "IDE0079:Remove unnecessary suppression", Justification = "False positive")]
     public sealed class Asmdef
     {
+        private readonly Project _project;
+        private readonly string _assetsFolder;
+
         public Asmdef(Project project)
         {
-            var properties = project.AllEvaluatedProperties;
+            _project = project;
+            _assetsFolder = _project
+                .AllEvaluatedProperties
+                .Single(p => p.Name == "AssetsFolder")
+                .EvaluatedValue;
 
-            Name = GetName(project.FullPath);
-            References = GetReferences(project.AllEvaluatedItems);
+            var properties = _project.AllEvaluatedProperties;
+
+            Name = GetName(_project.FullPath);
+            References = GetReferences();
             IncludePlatforms = GetCollectionProperty(properties, nameof(IncludePlatforms));
             ExcludePlatforms = GetCollectionProperty(properties, nameof(ExcludePlatforms));
             AllowUnsafeCode = GetBoolProperty(properties, "AllowUnsafeBlocks", false);
-            OverrideReferences = false;
-            PrecompiledReferences = ImmutableArray<string>.Empty;
+            OverrideReferences = true;
+            PrecompiledReferences = GetPrecompiledReferences();
             AutoReferenced = GetBoolProperty(properties, nameof(AutoReferenced), true);
             DefineConstraints = GetCollectionProperty(properties, nameof(DefineConstraints));
-            VersionDefines = GetCollectionProperty(properties, nameof(VersionDefines));
+            VersionDefines = ImmutableArray<string>.Empty;
             NoEngineReferences = GetBoolProperty(properties, nameof(NoEngineReferences), false);
         }
 
-        [JsonConstructor]
-        public Asmdef(
-            string name,
-            ImmutableArray<string> references,
-            ImmutableArray<string> includePlatforms,
-            ImmutableArray<string> excludePlatforms,
-            bool allowUnsafeCode,
-            bool overrideReferences,
-            ImmutableArray<string> precompiledReferences,
-            bool autoReferenced,
-            ImmutableArray<string> defineConstraints,
-            ImmutableArray<string> versionDefines,
-            bool noEngineReferences)
-        {
-            Name = name;
-            References = references;
-            IncludePlatforms = includePlatforms;
-            ExcludePlatforms = excludePlatforms;
-            AllowUnsafeCode = allowUnsafeCode;
-            OverrideReferences = overrideReferences;
-            PrecompiledReferences = precompiledReferences;
-            AutoReferenced = autoReferenced;
-            DefineConstraints = defineConstraints;
-            VersionDefines = versionDefines;
-            NoEngineReferences = noEngineReferences;
-        }
+        // ReSharper disable MemberCanBePrivate.Global
+        // ReSharper disable UnusedAutoPropertyAccessor.Global
 
         public string Name { get; }
         public ImmutableArray<string> References { get; }
@@ -69,7 +51,62 @@ namespace CsprojToAsmdef.Cli.Domain
         public ImmutableArray<string> VersionDefines { get; }
         public bool NoEngineReferences { get; }
 
-        public string CreateFileName() => Name + ".asmdef";
+        // ReSharper enable UnusedAutoPropertyAccessor.Global
+        // ReSharper enable MemberCanBePrivate.Global
+
+        public string GetFilePath() =>
+            Path.GetFullPath(
+                Path.Combine(
+                    _project.DirectoryPath,
+                    Name + ".asmdef"));
+
+        public string GetNuGetFolder() => Path.GetFullPath(Path.Combine(_assetsFolder, "NuGet"));
+
+        public string GetOutputDirectory()
+        {
+            // There must be a better way for this.
+            var outputDirectoryFromProject = _project
+                .AllEvaluatedProperties
+                .Where(p => p.Name == "OutputPath")
+                .Select(p => p.EvaluatedValue)
+                .Distinct()
+                .MaxBy(v => v.Length)
+                .First();
+
+            return Path.GetFullPath(Path.Combine(_project.DirectoryPath, outputDirectoryFromProject));
+        }
+
+        public ImmutableArray<string> GetFilesToCopy()
+        {
+            var outputDirectory = GetOutputDirectory();
+            var projectName = Path.GetFileNameWithoutExtension(_project.FullPath);
+            var referencedAsmdefs = References.Select(r => $"{r}.dll").ToImmutableArray();
+
+            return Directory
+                .EnumerateFiles(outputDirectory, "*", SearchOption.AllDirectories)
+                .Where(f =>
+                {
+                    var fileName = Path.GetFileName(f);
+
+                    return fileName != $"{projectName}.dll"
+                           && !f.EndsWith(".pdb")
+                           && !f.EndsWith(".deps.json")
+                           && !f.EndsWith(".dll.RoslynCA.json")
+                           && !referencedAsmdefs.Contains(fileName);
+                })
+                .ToImmutableArray();
+        }
+
+        public string CreateJson() =>
+            JsonSerializer
+                .Serialize(
+                    this,
+                    new JsonSerializerOptions(JsonSerializerDefaults.General)
+                    {
+                        PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
+                        WriteIndented = true,
+                    })
+                .Replace("  ", "    ");
 
         private static string GetName(string projectPath) => Path.GetFileNameWithoutExtension(projectPath);
 
@@ -92,23 +129,40 @@ namespace CsprojToAsmdef.Cli.Domain
                 : defaultValue;
         }
 
-        private static ImmutableArray<string> GetReferences(IEnumerable<ProjectItem> items) =>
-            items
-                .Where(i => i.ItemType == "Reference")
+        private ImmutableArray<string> GetReferences()
+        {
+            var upmReferences = _project.AllEvaluatedItems
+                .Where(i => i.ItemType is "Reference")
                 .Select(i => i.EvaluatedInclude)
-                .Where(i => i.Contains("ScriptAssemblies"))
-                .Select(i => Path.GetFileNameWithoutExtension(i)!)
+                .Where(i => i.Contains("ScriptAssemblies"));
+
+            var internalProjectReferences = _project.AllEvaluatedItems
+                .Where(i => i.ItemType is "ProjectReference")
+                .Select(i => i.EvaluatedInclude)
+                .Select(i => Path.IsPathRooted(i)
+                    ? i
+                    : Path.GetFullPath(Path.Combine(_project.DirectoryPath, i)))
+                .Where(i => i.StartsWith(_assetsFolder))
                 .ToImmutableArray();
 
-        public string CreateJson() =>
-            JsonSerializer
-                .Serialize(
-                    this,
-                    new JsonSerializerOptions(JsonSerializerDefaults.General)
-                    {
-                        PropertyNamingPolicy = JsonNamingPolicy.CamelCase,
-                        WriteIndented = true,
-                    })
-                .Replace("  ", "    ");
+            var transitiveAsmdefReferences = internalProjectReferences
+                .SelectMany(p => new Asmdef(new Project(p)).References);
+
+            return Enumerable.Empty<string>()
+                .Concat(upmReferences)
+                .Concat(internalProjectReferences)
+                .Select(i => Path.GetFileNameWithoutExtension(i)!)
+                .Concat(transitiveAsmdefReferences)
+                .Distinct()
+                .OrderBy(i => i)
+                .ToImmutableArray();
+        }
+
+        private ImmutableArray<string> GetPrecompiledReferences() =>
+            GetFilesToCopy()
+                .Where(f => Path.GetExtension(f) != "dll")
+                .Select(f => Path.GetFileName(f)!)
+                .OrderBy(f => f)
+                .ToImmutableArray();
     }
 }